Added User CRUD

+++ Backend/Controllers/blogControllers.js
@@ -1,8 +1,9 @@
+const mongoose = require ('mongoose');
 const Blog = require('../Models/Blog');
 
 const getAllBlogs =async (req, res) => {
     try {
-        const blogs = await Blog.find().createdAt( -1 );
+        const blogs = await Blog.find().sort({ createdAt: -1 });
         if (blogs){
             res.status(200).json(blogs);
         }
@@ -12,8 +13,8 @@ const getAllBlogs =async (req, res) => {
 }
 
 const getBlog =async (req, res) => {
+    const { id } = req.params;
     try{
-        const id = get.params;
         const blog = await Blog.findById(id);
         if(blog){
             res.status(200).json(blog);
@@ -24,23 +25,34 @@ const getBlog =async (req, res) => {
 }
 
 const createBlog =async (req, res) => {
+    const { title, snippet, body } = req.body;
+    const userID = "321456987"
+    //console.log({ userID, title, snippet, body});
     try{
-        const { title, snippet, body } = req.body;
-        const addedBlog = await Blog.save({ title, snippet, body});
+        const addedBlog = await Blog.create({ userID, title, snippet, body});
         if (addedBlog){
             res.status(200).json(addedBlog);
         }
+
     }catch (err){
         res.status(400).json(err);
     }
 }
 
 const updateBlog =async (req, res) => {
+    const { id } = req.params;
+    const  { title, snippet, body} = req.body;
+
     try{
-        const id = get.params;
-        const updatedBlog = await Blog.findByIdAndUpdate(id);
+        
+        const updatedBlog = await Blog.findByIdAndUpdate(id, { title, snippet, body}, { new:true });
         if (updatedBlog){
             res.status(200).json(updatedBlog);
+        }else{
+            if (!mongoose.Types.ObjectId.isValid(id)) {
+                return res.status(400).json({error: 'No such Blog'})
+            }
+            res.status(404).json({message:"Blog Not Found"});
         }
     }catch (err){
         res.status(400).json(err);
@@ -48,14 +60,20 @@ const updateBlog =async (req, res) => {
 }
 
 const deleteBlog =async (req, res) => {
-    try{
-        const id = get.params;
+    const { id } = req.params;
+    try{    
+        
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({error: 'No such workout'})
+        }
+
         const deletedBlog = await Blog.findByIdAndDelete(id);
         if (deletedBlog){
             res.status(200).json({"msg":"blog deleted"})
         }
+
     }catch(err){
-        res.status(404).json(err);
+        res.status(404).json({err:err.message});
     }
 }
